refactor: ReservationServiceImpl 매퍼 패턴 적용
