+++ reservation-service/src/main/java/com/carenest/business/reservationservice/application/service/ReservationServiceImpl.java
@@ -3,13 +3,14 @@ package com.carenest.business.reservationservice.application.service;
 import com.carenest.business.reservationservice.application.dto.request.ReservationCreateRequest;
 import com.carenest.business.reservationservice.application.dto.request.ReservationSearchRequest;
 import com.carenest.business.reservationservice.application.dto.request.ReservationUpdateRequest;
-import com.carenest.business.reservationservice.application.dto.response.ReservationResponse;
 import com.carenest.business.reservationservice.domain.model.Reservation;
 import com.carenest.business.reservationservice.domain.model.ReservationStatus;
 import com.carenest.business.reservationservice.domain.model.PaymentStatus;
 import com.carenest.business.reservationservice.domain.repository.ReservationRepository;
 import com.carenest.business.reservationservice.domain.service.ReservationDomainService;
 import com.carenest.business.reservationservice.exception.*;
+import com.carenest.business.reservationservice.presentation.dto.mapper.ReservationMapper;
+import com.carenest.business.reservationservice.presentation.dto.response.ReservationResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -25,6 +26,7 @@ public class ReservationServiceImpl implements ReservationService {
 
     private final ReservationRepository reservationRepository;
     private final ReservationDomainService reservationDomainService;
+    private final ReservationMapper reservationMapper;
 
     @Override
     @Transactional
@@ -55,7 +57,7 @@ public class ReservationServiceImpl implements ReservationService {
 
         reservationDomainService.createReservationHistory(savedReservation);
 
-        return new ReservationResponse(savedReservation);
+        return reservationMapper.toDto(savedReservation);
     }
 
     @Override
@@ -64,7 +66,7 @@ public class ReservationServiceImpl implements ReservationService {
         Reservation reservation = reservationRepository.findById(reservationId)
                 .orElseThrow(ReservationNotFoundException::new);
 
-        return new ReservationResponse(reservation);
+        return reservationMapper.toDto(reservation);
     }
 
     @Override
@@ -78,7 +80,7 @@ public class ReservationServiceImpl implements ReservationService {
         }
 
         Page<Reservation> reservations = reservationRepository.findByStartedAtBetween(startDate, endDate, pageable);
-        return reservations.map(ReservationResponse::new);
+        return reservations.map(reservationMapper::toDto);
     }
 
     @Override
@@ -93,7 +95,7 @@ public class ReservationServiceImpl implements ReservationService {
                 request.getStatus(),
                 pageable
         );
-        return reservations.map(ReservationResponse::new);
+        return reservations.map(reservationMapper::toDto);
     }
 
     @Override
@@ -112,14 +114,14 @@ public class ReservationServiceImpl implements ReservationService {
                 userId, startDate, endDate, pageable);
 
         if (!reservationsByGuardian.isEmpty()) {
-            return reservationsByGuardian.map(ReservationResponse::new);
+            return reservationsByGuardian.map(reservationMapper::toDto);
         }
 
         // 간병인 ID로 조회
         Page<Reservation> reservationsByCaregiver = reservationRepository.findByCaregiverIdAndStartedAtBetween(
                 userId, startDate, endDate, pageable);
 
-        return reservationsByCaregiver.map(ReservationResponse::new);
+        return reservationsByCaregiver.map(reservationMapper::toDto);
     }
 
     @Override
@@ -154,7 +156,7 @@ public class ReservationServiceImpl implements ReservationService {
         Reservation updatedReservation = reservationRepository.save(reservation);
         reservationDomainService.createReservationHistory(updatedReservation);
 
-        return new ReservationResponse(updatedReservation);
+        return reservationMapper.toDto(updatedReservation);
     }
 
     @Override
@@ -172,7 +174,7 @@ public class ReservationServiceImpl implements ReservationService {
 
         reservationDomainService.createReservationHistory(reservation);
 
-        return new ReservationResponse(reservation);
+        return reservationMapper.toDto(reservation);
     }
 
     @Override
@@ -190,7 +192,7 @@ public class ReservationServiceImpl implements ReservationService {
 
         reservationDomainService.createReservationHistory(reservation);
 
-        return new ReservationResponse(reservation);
+        return reservationMapper.toDto(reservation);
     }
 
     @Override
@@ -214,7 +216,7 @@ public class ReservationServiceImpl implements ReservationService {
 
         reservationDomainService.createReservationHistory(reservation);
 
-        return new ReservationResponse(reservation);
+        return reservationMapper.toDto(reservation);
     }
 
     @Override
@@ -228,7 +230,7 @@ public class ReservationServiceImpl implements ReservationService {
         }
 
         Page<Reservation> reservations = reservationRepository.findByStartedAtBetween(startDate, endDate, pageable);
-        return reservations.map(ReservationResponse::new);
+        return reservations.map(reservationMapper::toDto);
     }
 
     @Override
@@ -246,21 +248,21 @@ public class ReservationServiceImpl implements ReservationService {
                 userId, startDate, endDate, pageable);
 
         if (!reservationsByGuardian.isEmpty()) {
-            return reservationsByGuardian.map(ReservationResponse::new);
+            return reservationsByGuardian.map(reservationMapper::toDto);
         }
 
         // 간병인 ID로 조회
         Page<Reservation> reservationsByCaregiver = reservationRepository.findByCaregiverIdAndStartedAtBetween(
                 userId, startDate, endDate, pageable);
 
-        return reservationsByCaregiver.map(ReservationResponse::new);
+        return reservationsByCaregiver.map(reservationMapper::toDto);
     }
 
     @Override
     @Transactional(readOnly = true)
     public Page<ReservationResponse> getReservationsByStatus(ReservationStatus status, Pageable pageable) {
         Page<Reservation> reservations = reservationRepository.findByStatus(status, pageable);
-        return reservations.map(ReservationResponse::new);
+        return reservations.map(reservationMapper::toDto);
     }
 
     @Override
@@ -278,7 +280,7 @@ public class ReservationServiceImpl implements ReservationService {
 
         reservationDomainService.createReservationHistory(reservation);
 
-        return new ReservationResponse(reservation);
+        return reservationMapper.toDto(reservation);
     }
 
     @Override
@@ -297,6 +299,6 @@ public class ReservationServiceImpl implements ReservationService {
         Reservation updatedReservation = reservationRepository.save(reservation);
         reservationDomainService.createReservationHistory(updatedReservation);
 
-        return new ReservationResponse(updatedReservation);
+        return reservationMapper.toDto(updatedReservation);
     }
 }
